+++ aiosenseme/__init__.py
@@ -1,7 +1,9 @@
 """aiosenseme library."""
 
-from .device import SensemeDevice, SensemeFan, SensemeLight
-from .discovery import SensemeDiscovery, discover, discover_any
+from .device import (SensemeDevice, SensemeFan, SensemeLight,
+                     async_get_device_by_device_info,
+                     async_get_device_by_ip_address)
+from .discovery import SensemeDiscovery, discover, discover_all, discover_any
 from .version import __version__
 
 __all__ = [
@@ -9,8 +11,11 @@ __all__ = [
     "SensemeFan",
     "SensemeLight",
     "SensemeDiscovery",
+    "discover_all",
     "discover_any",
     "discover",
+    "async_get_device_by_ip_address",
+    "async_get_device_by_device_info",
 ]
 
 # flake8: noqa